backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.211.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
elasticsearch==8.11.0
aiohttp==3.9.0
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.25.1
authlib==1.3.0
itsdangerous==2.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
faker==20.0.3
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
networkx==3.2.1
biopython==1.81
bibtexparser==1.4.0
lxml==4.9.3
requests==2.31.0
scipy==1.11.4
joblib==1.3.2
gunicorn==21.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
sentence-transformers==2.3.1
tqdm==4.66.1
xgboost==2.0.0
treeinterpreter==0.2.3
openai>=1.70.0  # Required for GPT-5 support and reasoning_effort parameter
umap-learn==0.5.5
hdbscan==0.8.40
isodate==0.7.2
markdown==3.5.1
beautifulsoup4==4.12.2
html2text==2020.1.16
PyGithub==2.1.1
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0